requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.1
faiss-cpu==1.7.4
pypdf==4.2.0
python-dotenv==1.0.1
langchain-google-genai==0.1.3
google-generativeai==0.5.2
httpx==0.26.0
pydantic-settings==2.1.0
pydantic==2.11.0
streamlit==1.35.0